Contrato guardando datos, eliminacion falta probar, estructuras asociadas creadas sin vista

+++ ejb/src/main/java/com/uy/nos/financiarte/controller/RegistroContrato.java
@@ -12,8 +12,14 @@ import javax.inject.Inject;
 import javax.inject.Named;
 import javax.persistence.EntityManager;
 
+import com.uy.nos.financiarte.data.DevolucionListProducer;
+import com.uy.nos.financiarte.data.FacturaListProducer;
+import com.uy.nos.financiarte.data.NotaCreditoListProducer;
+import com.uy.nos.financiarte.data.PagoClienteListProducer;
+import com.uy.nos.financiarte.data.PagoMedioPagoListProducer;
 import com.uy.nos.financiarte.model.Cliente;
 import com.uy.nos.financiarte.model.Contrato;
+import com.uy.nos.financiarte.model.Interes;
 import com.uy.nos.financiarte.model.Proveedor;
 
 
@@ -27,26 +33,46 @@ public class RegistroContrato {
 
 	   @Inject
 	   private EntityManager em;
+	   
+	   @Inject
+	   private DevolucionListProducer dlp;
+	   
+	   @Inject
+	   private PagoClienteListProducer pclp;
+	   
+	   @Inject
+	   private FacturaListProducer flp;
+	   
+	   @Inject
+	   private NotaCreditoListProducer nclp;
+	   
+	   @Inject
+	   private PagoMedioPagoListProducer pmplp;
 
 	   @Inject
 	   private Event<Contrato> contratoEventSrc;
 
 	   private Contrato newContrato;
+	   private boolean tieneDatos = false;
 
 	   @Produces
 	   @Named
 	   public Contrato getNewContrato() {
 	      return newContrato;
 	   }
-
-	   public void registro(Cliente clienteSeleccionado, Proveedor proveedorSeleccionado) throws Exception {
+	   
+	   public void registro(Cliente cliente, Proveedor proveedor, long montoPrestamo, int diasInteres, long pagoMinimo, int plazoPago, Interes interes) throws Exception {
 	      log.info("Registro " + newContrato.getId());
 	      Calendar today = Calendar.getInstance();
 	      today.set(Calendar.HOUR_OF_DAY, 0);
 	      newContrato.setFecha(today);
-	      newContrato.setCliente(clienteSeleccionado);
-	      newContrato.setProveedor(proveedorSeleccionado);
-	      newContrato.setInteres(newContrato.getInteres());
+	      newContrato.setCliente(cliente);
+	      newContrato.setProveedor(proveedor);
+	      newContrato.setInteres(interes);
+	      newContrato.setMontoPrestamo(montoPrestamo);
+	      newContrato.setDiasInteres(diasInteres);
+	      newContrato.setPagoMinimo(pagoMinimo);
+	      newContrato.setPlazoPago(plazoPago);
 	      em.persist(newContrato);
 	      contratoEventSrc.fire(newContrato);
 	      initNewComercio();
@@ -59,9 +85,12 @@ public class RegistroContrato {
 	   
 	   public void eliminar(Long id) throws Exception {
 		   log.info("Elimino " + id);
-		   Contrato contrato = em.find(Contrato.class, id);
-		   em.remove(contrato);
-		   contratoEventSrc.fire(newContrato);
+		   contratoTieneDatos(id);
+		   if (tieneDatos == false){
+			   Contrato contrato = em.find(Contrato.class, id);
+			   em.remove(contrato);
+			   contratoEventSrc.fire(newContrato);
+		   }
 	   }
 
 	   public Contrato buscar(Long id) throws Exception {
@@ -73,6 +102,25 @@ public class RegistroContrato {
 	   @PostConstruct
 	   public void initNewComercio() {
 		   newContrato = new Contrato();
+	   }
+	   
+	   public void contratoTieneDatos(Long idContrato){
+		   
+		   if(!dlp.getDevolucionPorContrato(idContrato).isEmpty()){
+			   tieneDatos = true;
+		   }
+		   if(!pclp.getPagoClientePorContrato(idContrato).isEmpty()){
+			   tieneDatos = true;
+		   }
+		   if(!flp.getFacturaPorContrato(idContrato).isEmpty()){
+			   tieneDatos = true;
+		   }
+		   if(!nclp.getNotaCreditoPorContrato(idContrato).isEmpty()){
+			   tieneDatos = true;
+		   }
+		   if(!pmplp.getPagoMedioPagoPorContrato(idContrato).isEmpty()){
+			   tieneDatos = true;
+		   }
 	   }
 	   
 }
